git-svn-id: http://svn.osgeo.org/qgis/trunk/qgis@4409 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ plugins/georeferencer/datapointacetate.h
@@ -17,15 +17,16 @@ public:
 				     const QgsPoint& mapCoords)
   // UGLY!
     : QgsAcetateObject(const_cast<QgsPoint&>(pixelCoords)),
-      mCoords(mapCoords) {
+      mCoords(mapCoords) 
+  {
     
   }
 
   virtual void draw(QPainter* painter, QgsMapToPixel* cXf = 0) {
     if (cXf) {
       QgsPoint pixelOrigin = cXf->transform(origin());
-      int x(pixelOrigin.x());
-      int y(pixelOrigin.y());
+      int x(static_cast<int>(pixelOrigin.x()));
+      int y(static_cast<int>(pixelOrigin.y()));
       QFont font;
       painter->setFont(QFont("helvetica", 9));
       painter->setPen(Qt::black);
